feat(prompts): paginate Top Sources tables like the citations page
Replace the show-all toggle with the citations-style TablePager (100
rows per page, continuous rank across pages). TablePager and
usePagination moved from the citations page into components/table-pager
so both pages share one implementation; the reset-on-key-change logic
now uses the adjust-state-during-render pattern instead of a ref write,
which the react-hooks/refs lint rule rejects.
